+++ docs/site/app/favicon.ico
